# argus-backend/api/openapi.yaml
openapi: 3.1.0
info:
  title: Argus AI — Event Collector API
  version: 2.1.0
  description: |
    REST API for the Argus AI proctoring platform.

    ## Authentication

    All endpoints (except `POST /api/v1/auth/login`) require a Bearer JWT in the
    `Authorization` header:

    ```
    Authorization: Bearer <token>
    ```

    Tokens are obtained via the login endpoint and expire after 24 hours.

    ## Rate Limits

    | Scope       | Limit             |
    |-------------|-------------------|
    | Global      | 5 000 req/s       |
    | Per-IP      | 100 req/s         |
    | Login       | 5 req/s (Nginx)   |

    Rate-limit responses return HTTP 429 with a `Retry-After` header.

  contact:
    name: Argus AI Engineering
    url: https://argus.ai
servers:
  - url: https://api.argus.ai
    description: Production
  - url: http://localhost:8080
    description: Local development
...
paths:
...
  /api/v1/media/token:
    post:
      tags: [Media]
      summary: Get LiveKit WebRTC token
      description: |
        Generates a signed LiveKit access token for joining a proctoring room.

        `role` values:
        - `proctor` — can subscribe to all tracks, publish audio
        - `student` — can publish camera/microphone/screen, subscribe to nothing
      operationId: getMediaToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/MediaTokenRequest'
      responses:
        '200':
          description: LiveKit token
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/MediaTokenResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/media/rooms:
    get:
      tags: [Media]
      summary: List active LiveKit rooms
      operationId: listMediaRooms
      parameters:
        - $ref: '#/components/parameters/OrgIdQuery'
      responses:
        '200':
          description: Room list
          content:
            application/json:
              schema:
                type: array
                items:
                  type: object
                  properties:
                    room:
                      type: string
                    participantCount:
                      type: integer
        '401':
          $ref: '#/components/responses/Unauthorized'
  /api/v1/media/recordings/start:
    post:
      tags: [Media]
      summary: Start LiveKit egress recording
      description: |
        Starts a LiveKit Track Composite Egress for selected student tracks.

        The backend writes an initial row to PostgreSQL table `livekit_recordings`
        keyed by the returned `egressId`. Recording output is MP4 at 720p, 15 FPS,
        written to MinIO/S3 under:

        `content/recordings/{roomName}/{sessionId}-{studentId}.mp4`

        If PostgreSQL persistence fails after LiveKit starts, backend attempts to
        stop the new egress immediately to avoid orphan recordings.
      operationId: startMediaRecording
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StartRecordingRequest'
      responses:
        '202':
          description: Recording accepted and egress started
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          description: LiveKit failed to start the egress
  /api/v1/media/recordings/stop:
    post:
      tags: [Media]
      summary: Stop LiveKit egress recording
      description: |
        Stops a running LiveKit egress by `egressId` and marks the local
        metadata row as `EGRESS_ENDING`. The final file URL and final status are
        written later by the signed LiveKit `egress_ended` webhook.
      operationId: stopMediaRecording
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/StopRecordingRequest'
      responses:
        '200':
          description: Stop request accepted by LiveKit
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/RecordingResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '401':
          $ref: '#/components/responses/Unauthorized'
        '500':
          $ref: '#/components/responses/InternalServerError'
        '502':
          description: LiveKit failed to stop the egress
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT

  # ── Parameters ─────────────────────────────────────────────────────────────
  parameters:
...
    OrgIdQuery:
      name: orgId
      in: query
      schema:
        type: string
      description: Filter results to a specific organization. Super admins may omit for all orgs.
...
  responses:
    Unauthorized:
      description: Missing or invalid Bearer token
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "unauthorized"
...
    BadRequest:
      description: Invalid request body or parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "invalid request body"
    InternalServerError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ErrorResponse'
          example:
            error: "internal server error"
...
  schemas:
...
    ErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          type: string
...
    MediaTokenRequest:
      type: object
      required: [sessionId, studentId, examId, role]
      properties:
        sessionId:
          type: string
        studentId:
          type: string
        examId:
          type: string
        role:
          type: string
          enum: [proctor, student]
    MediaTokenResponse:
      type: object
      properties:
        token:
          type: string
          description: Signed LiveKit access token (JWT)
        wsUrl:
          type: string
          description: LiveKit WebSocket URL (e.g. `wss://livekit.argus.ai`)
        room:
          type: string
          description: Room name derived from sessionId
    StartRecordingRequest:
      type: object
      required: [sessionId, studentId, videoTrackId]
      properties:
        sessionId:
          type: string
          description: Argus proctoring session ID
        studentId:
          type: string
          description: Student/user ID associated with the recording
        roomName:
          type: string
          description: Optional LiveKit room name. Defaults to `argus-session-{sessionId}`.
        videoTrackId:
          type: string
          description: LiveKit video track ID to record
        audioTrackId:
          type: string
          description: Optional LiveKit audio track ID to include
    StopRecordingRequest:
      type: object
      required: [egressId]
      properties:
        egressId:
          type: string
          description: LiveKit Egress ID returned by the start endpoint
    RecordingResponse:
      type: object
      properties:
        egressId:
          type: string
          description: LiveKit Egress ID
        status:
          type: string
          description: Local recording status such as `EGRESS_ACTIVE` or `EGRESS_ENDING`

    # ── Ingest (Chunk) ───────────────────────────────────────────────────────
...
security:
  - BearerAuth: []